refactor: merge the sample agent so we have one email inbox management agent
